refactor: isolate admin composition root
