+++ crates/dcc-mcp-gateway/src/gateway/admin/analytics_tests.rs
@@ -11,7 +11,7 @@ use serde_json::Value;
 use tokio::sync::{RwLock, broadcast, watch};
 use tower::ServiceExt;
 
-use crate::gateway::admin::router::{build_admin_router, build_v1_debug_router};
+use crate::gateway::admin::application::router::{build_admin_router, build_v1_debug_router};
 use crate::gateway::admin::state::{AdminAuditRecord, AdminState, AuditLog};
 use crate::gateway::admin::trace::TokenTelemetry;
 use crate::gateway::state::GatewayState;
