+++ crates/dcc-mcp-gateway/src/gateway/admin/application/handlers.rs
@@ -14,10 +14,8 @@ use serde_json::{Value, json};
 
 use super::debug_response::{DebugListQuery, debug_response};
 use super::events::contend_event_to_admin_row;
+use super::issue_report::{IssueReportMode, issue_report_filename, issue_report_json};
 use super::skill_reload::reload_skill_paths_and_refresh_backends;
-use crate::gateway::admin::issue_report::{
-    IssueReportMode, issue_report_filename, issue_report_json,
-};
 use crate::gateway::admin::state::{AdminAuditRecord, AdminState};
 use crate::gateway::admin::trace::{AgentContext, DispatchTrace};
 use crate::gateway::capability::RefreshReason;
