+++ crates/dcc-mcp-gateway/src/gateway/admin/basic_endpoint_tests.rs
@@ -12,7 +12,7 @@ mod endpoint_contracts {
     use tokio::sync::{RwLock, broadcast, watch};
     use tower::ServiceExt;
 
-    use crate::gateway::admin::router::build_admin_router;
+    use crate::gateway::admin::application::router::build_admin_router;
     use crate::gateway::admin::state::AdminState;
     use crate::gateway::state::GatewayState;
     use dcc_mcp_transport::discovery::file_registry::FileRegistry;
