+++ crates/dcc-mcp-gateway/src/gateway/admin/infra/wecom_url.rs
