+++ crates/dcc-mcp-gateway/src/gateway/admin/marketplace_tests.rs
@@ -14,7 +14,7 @@ use serde_json::{Value, json};
 use tokio::sync::{RwLock, broadcast, watch};
 use tower::ServiceExt;
 
-use crate::gateway::admin::router::build_admin_router;
+use crate::gateway::admin::application::router::build_admin_router;
 use crate::gateway::admin::state::AdminState;
 use crate::gateway::state::GatewayState;
 use dcc_mcp_transport::discovery::file_registry::FileRegistry;
