+++ crates/dcc-mcp-gateway/src/gateway/admin/tests.rs
@@ -17,7 +17,7 @@ pub(in crate::gateway::admin) mod admin_tests {
 
     use dcc_mcp_gateway_core::capability_naming::instance_short;
 
-    use crate::gateway::admin::router::{build_admin_router, build_v1_debug_router};
+    use crate::gateway::admin::application::router::{build_admin_router, build_v1_debug_router};
     use crate::gateway::admin::state::{AdminAuditRecord, AdminState, AuditLog};
     use crate::gateway::admin::trace::{AgentContextTrust, TokenTelemetry};
     use crate::gateway::router::build_gateway_router_with_admin;
